SMP to startup all APs

+++ include/drivers/acpi/madt.h
@@ -66,7 +66,7 @@ struct MADT {
 // Our own struct that contains usable data parsed from MADT.
 struct MADT_info {
     void* local_apic_addr;
-    uint16_t processor_count;
+    uint8_t processor_count;
     uint8_t processor_ids[MAX_CPUS_SUPPORTED];
     uint8_t local_APIC_ids[MAX_CPUS_SUPPORTED];
     uint16_t io_apic_count;
